requirements.txt
requests
pyjwt
Werkzeug
flask_cors